Split off CRT related stuff into a plugin
