+++ src/main.rs
@@ -19,6 +19,7 @@ pub mod ui;
 pub mod assets;
 pub mod inventory;
 pub mod editor;
+pub mod crt;
 
 pub fn main() {
     App::new()
@@ -26,7 +27,6 @@ pub fn main() {
         .add_plugins(DefaultPlugins)
         .insert_resource(RapierConfiguration::default())
         .insert_resource(Selected { entity: None })
-        .insert_resource(ScreenActivated { entity: None })
         .add_plugin(RapierPhysicsPlugin::<NoUserData>::default())
         //.add_plugin(RapierDebugRenderPlugin::default())
         .add_plugin(FpsControllerPlugin)
@@ -35,12 +35,12 @@ pub fn main() {
         .add_plugin(crate::ui::UiPlugin)
         .add_plugin(crate::inventory::InventoryPlugin)
         .add_plugin(crate::key_translator::KeyTranslatorPlugin)
+        .add_plugin(crate::crt::CrtPlugin)
         //.add_plugin(Sprite3dPlugin)
         //.add_plugin(crate::camera::PlayerPlugin)
         .add_system(manage_cursor)
         .add_system(interaction_glow)
         .add_system(interact)
-        .add_system(run_editor)
         .add_startup_system(setup)
         //.add_system(movement)
         .run();
@@ -56,11 +56,6 @@ pub struct Item {
 
 #[derive(Resource)]
 pub struct Selected {
-    entity: Option<Entity>,
-}
-
-#[derive(Resource)]
-pub struct ScreenActivated {
     entity: Option<Entity>,
 }
 
@@ -137,41 +132,6 @@ fn setup(
             color: Color::rgba(1.0, 1.0, 1.0, 1.0),
         }));
 
-    {
-        use bevy::render::render_resource::*;
-
-        let mut image = Image::new_fill(
-            Extent3d { width: 1360, height: 768, depth_or_array_layers: 1 },
-            TextureDimension::D2,
-            &[255u8, 0u8, 255u8, 255u8],
-            TextureFormat::Bgra8UnormSrgb);
-
-        let material_handle = materials.add(StandardMaterial {
-            base_color_texture: Some(images.add(image)),
-            // reflectance: 0.02,
-            unlit: false,
-            ..default()
-        });
-
-        commands.spawn_bundle((
-            crate::editor::Screen::new(editor::Editor::new()),
-            PbrBundle {
-                mesh: meshes.add(Mesh::from(shape::Plane { size: 0.25 })),
-                material: material_handle,
-                transform: Transform::from_xyz(1.0, 0.75, 1.0),
-                    //.looking_at(Vec3::new(1.5, 1.5, 1.5), Vec3::new(0.0, 1.0, 0.0)),
-                ..default()
-            },
-            Interactable,
-            Collider::cuboid(0.125, 0.01, 0.125),
-            outlines.add(OutlineMaterial {
-                width: 0.,
-                color: Color::rgba(1.0, 1.0, 1.0, 1.0),
-            }),
-        ));
-
-    }
-
     commands.spawn_bundle(PointLightBundle {
         point_light: PointLight {
             intensity: 1500.0,
@@ -181,56 +141,6 @@ fn setup(
         transform: Transform::from_xyz(4.0, 8.0, 4.0),
         ..default()
     });
-}
-
-pub fn run_editor(
-    mut images: ResMut<Assets<Image>>,
-    mut materials: ResMut<Assets<StandardMaterial>>,
-    mut screens: Query<(&mut crate::editor::Screen, &Handle<StandardMaterial>)>,
-    screen_activated: Res<ScreenActivated>,
-    mut keyboard_events: EventReader<TranslatedKey>,
-) {
-    if let Some(entity) = screen_activated.entity {
-        let (mut screen, material_handle) = screens.get_mut(entity).unwrap();
-
-        let mut needs_rerender = false;
-
-        for key in keyboard_events.iter() {
-            if key.pressed {
-                screen.editor.process_keypress(key.key);
-                needs_rerender = true;
-            }
-        }
-
-        if !needs_rerender {
-            return;
-        }
-
-        screen.editor.refresh_screen();
-
-        let rasterized = screen.editor.rasterize().unwrap();
-
-        let image_handle =
-            materials.get_mut(material_handle).unwrap()
-            .base_color_texture.clone().unwrap();
-        let image: &mut Image = images.get_mut(&image_handle).unwrap();
-
-        {
-            let mut index = 0;
-            for y in 0 .. rasterized.height {
-                for x in 0 .. rasterized.width {
-                    let [a, r, g, b] =
-                        rasterized.get((rasterized.width - 1) - x, y).to_le_bytes();
-                    image.data[index + 0] = b;
-                    image.data[index + 1] = g;
-                    image.data[index + 2] = r;
-                    image.data[index + 3] = a;
-                    index += 4;
-                }
-            }
-        }
-
-    }
 }
 
 pub fn manage_cursor(
@@ -263,7 +173,7 @@ pub fn interact(
     items: Query<&Item>,
     mut inventories: Query<&mut Inventory, With<LogicalPlayer>>,
     screens: Query<&crate::editor::Screen>,
-    mut screen_activated: ResMut<ScreenActivated>,
+    mut screen_activated: ResMut<crate::crt::ScreenActivated>,
     mut controllers: Query<&mut FpsController>,
 ) {
     let mut inventory = inventories.single_mut();
